+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.3.0-alpha.1
+**Maintainer**: Sandro Aebischer <aebischer@gyselroth.com>\
+**Date**: Wed June 02 15:25:00 CET 2021
+
+### Features
+* Added option to define if interval refers to start or end of previous job
+
+
 ## 3.2.2
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
 **Date**: Thu Dec 05 09:38:31 CET 2019